more patches to main typings file so it properly exports require in the correct scope
