+++ webpack/common.js
@@ -1,5 +1,7 @@
 const _ = require("lodash");
+const fs = require("fs");
 const path = require("path");
+const shell = require("shelljs");
 const assert = require("assert");
 const debug = require("debug")("CdkWeb:Common");
 
@@ -48,4 +50,79 @@ class PathTracker {
   };
 }
 
-module.exports = { ...Constants, MakeSureReplaced, PathTracker };
+const getModules = _.memoize(() => {
+  const { stdout: folders } = shell.exec(`find -type d -maxdepth 1`, {
+    cwd: path.resolve(__ROOT, `node_modules/${"aws-cdk-lib"}`),
+    silent: true,
+  });
+  const paths = [
+    "fs",
+    "path",
+    "aws-sdk",
+    "constructs",
+    "aws-cdk-lib",
+    ...folders
+      .trim()
+      .split("\n")
+      .map((p) => p.replace("./", `${"aws-cdk-lib"}/`))
+      .filter((m) => m !== ".")
+      .filter((m) =>
+        _.chain(() => require.resolve(m))
+          .attempt()
+          .isError()
+      ),
+  ].filter((m) => {
+    try {
+      require(m);
+      debug("[x] %s", m);
+      return true;
+    } catch (err) {
+      debug("[ ] %s", m);
+      return false;
+    }
+  });
+  return paths;
+});
+
+const getAssets = _.memoize(() => {
+  const libCwd = path.resolve(__ROOT, "node_modules/aws-cdk-lib");
+  const findJsonCmd = "find -wholename './**/*.json' | awk '!/node/ && !/.vscode/ && !/jsii/'";
+  const { stdout: libAssets } = shell.exec(findJsonCmd, { silent: true, cwd: libCwd });
+  const findYamlCmd = "find -wholename './**/*.yaml' | awk '!/node/ && !/.vscode/ && !/jsii/'";
+  const cliCwd = path.resolve(__ROOT, "node_modules/aws-cdk");
+  const { stdout: cliAssets } = shell.exec(findYamlCmd, { silent: true, cwd: cliCwd });
+  const postProcess = (assets, cwd) =>
+    assets
+      .trim()
+      .split("\n")
+      .map((module) => ({
+        path: `/${path.basename(module)}`,
+        code: fs.readFileSync(path.resolve(cwd, module), {
+          encoding: "utf-8",
+        }),
+      }));
+  const assets = [...postProcess(cliAssets, cliCwd), ...postProcess(libAssets, libCwd)];
+  assets.push({
+    path: "/cdk.json",
+    code: JSON.stringify(
+      {
+        app: "index.js",
+        context: {
+          "@aws-cdk/aws-apigateway:usagePlanKeyOrderInsensitiveId": true,
+          "@aws-cdk/core:stackRelativeExports": true,
+          "@aws-cdk/aws-rds:lowercaseDbIdentifier": true,
+          "@aws-cdk/aws-lambda:recognizeVersionProps": true,
+          "@aws-cdk/aws-cloudfront:defaultSecurityPolicyTLSv1.2_2021": true,
+          "@aws-cdk-containers/ecs-service-extensions:enableDefaultLogDriver": true,
+          "@aws-cdk/aws-ec2:uniqueImdsv2TemplateName": true,
+          "@aws-cdk/core:target-partitions": ["aws", "aws-cn"],
+        },
+      },
+      null,
+      2
+    ),
+  });
+  return assets;
+});
+
+module.exports = { ...Constants, MakeSureReplaced, PathTracker, getAssets, getModules };
